# openapi.yaml
openapi: 3.1.0
info:
  title: HighFive! API
  version: 2.0.0
  description: >-
    Contrat backend v2, derive de `src/domain/**` (schemas zod), `src/api/*.ts`
    (appels reels du front) et `src/mocks/handlers/**` (regles metier
    appliquees cote mock). Voir `docs/v2/backend/SPEC.md` pour le detail des
    regles metier, du modele de persistance propose et des ecarts connus.
    Les schemas de composants sont generes depuis `src/domain/index.ts` par
    `pnpm export:schemas` (zod 4, `z.toJSONSchema`) dans `docs/v2/backend/schemas/`.
servers:
  - url: /api
    description: Prefixe fixe (voir `src/api/client.ts`), au-dessus de `VITE_API_URL`.
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: Fil des projets publics actifs (R-V1)
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
          description: Recherche texte (titre, accroche).
        - name: tags
          in: query
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
          description: Filtre additif par identifiants de tags (liste separee par virgules cote HTTP).
        - name: participation
          in: query
          schema: { $ref: "#/components/schemas/Participation" }
          description: >-
            Declare cote client (`ListProjectsQuery`) mais non applique par le
            mock actuel — a implementer cote backend, voir SPEC.md.
        - name: sort
          in: query
          schema: { type: string, enum: [recent, popular, relevant] }
          description: >-
            Declare cote client mais non applique par le mock actuel (tri fixe
            sur `lastActivityAt`) — a implementer cote backend, voir SPEC.md.
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Page de projets.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedProjectSummary" }
    post:
      operationId: createProject
      tags: [projects]
      summary: Creer un projet (etat initial `draft`)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectCreateInput" }
      responses:
        "201":
          description: Projet cree, porteur = personne connectee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Compte non `active` (suspendu).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Fiche d'un projet
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Projet complet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "403":
          description: Projet prive et lecteur hors equipe (R-V3).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Modifier titre/accroche/description/tags/besoins/visibilite/participation
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdateInput" }
      responses:
        "200":
          description: Projet mis a jour.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve porteur/co-porteur (R-PR1).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Supprimer definitivement (R-PR7, confirmation nominative)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectDeleteInput" }
      responses:
        "204":
          {
            description: "Projet supprime (differe 30 jours cote donnees, R-X3 ; voir SPEC.md).",
          }
        "400":
          description: Le titre saisi ne correspond pas (`confirmTitle`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve au porteur et a l'administration.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /projects/{slug}/announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: Annonces d'un projet (auteur resolu)
      security: []
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Annonces.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AnnouncementWithAuthor" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Publier une annonce (R-A1, notifie membres + highfiveurs R-A3)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementCreateInput" }
      responses:
        "201":
          description: Annonce publiee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve porteur/co-porteur (R-A1).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
  /announcements/{announcementId}/pin:
    post:
      operationId: pinAnnouncement
      tags: [announcements]
      summary: Epingler (R-A2 — depingle automatiquement l'ancienne)
      parameters:
        - $ref: "#/components/parameters/AnnouncementIdParam"
      responses:
        "200":
          description: Annonce epinglee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /announcements/{announcementId}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Supprimer une annonce
      parameters:
        - $ref: "#/components/parameters/AnnouncementIdParam"
      responses:
        "204": { description: Annonce supprimee. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------- Comments
...
  /admin/reports:
    get:
      operationId: listReports
      tags: [admin]
      summary: File de moderation (R-S2 — 3+ signalements en tete)
      parameters:
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Page de signalements, avec signaleur, apercu de la cible et compte de signalements similaires.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedReportSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/reports/{reportId}/resolve:
    post:
      operationId: resolveReport
      tags: [admin]
      summary: Traiter un signalement (R-S4 journalise)
      parameters:
        - $ref: "#/components/parameters/ReportIdParam"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReasonInput" }
      responses:
        "200":
          description: "Signalement `status: handled`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/reports/{reportId}/reject:
    post:
      operationId: rejectReport
      tags: [admin]
      summary: Rejeter un signalement (R-S4 journalise)
      parameters:
        - $ref: "#/components/parameters/ReportIdParam"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReasonInput" }
      responses:
        "200":
          description: "Signalement `status: rejected`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Statistiques de la plateforme (calculees en direct)
      responses:
        "200":
          description: Statistiques.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminStats" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: Comptes (mot de passe jamais renvoye)
      parameters:
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Page de comptes.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCurrentUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users/{userId}/suspend:
    post:
      operationId: suspendUser
      tags: [admin]
      summary: Suspendre un compte (journalise, notifie R-RP3)
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReasonInput" }
      responses:
        "204": { description: "Compte `accountStatus: suspended`." }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/users/{userId}/reactivate:
    post:
      operationId: reactivateUser
      tags: [admin]
      summary: Lever une suspension (journalise, notifie R-RP3)
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReasonInput" }
      responses:
        "204": { description: "Compte `accountStatus: active`." }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/projects:
    get:
      operationId: listAdminProjects
      tags: [admin]
      summary: Tous les projets (tout etat)
      parameters:
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Page de projets.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedProjectSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/projects/{slug}:
    delete:
      operationId: adminDeleteProject
      tags: [admin]
      summary: >-
        Supprimer un projet (R-PR7, journalise). Non appele par l'interface
        actuelle, qui reutilise `DELETE /projects/{slug}` avec confirmation
        nominative — route conservee au contrat, voir SPEC.md.
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "204": { description: Projet supprime. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/tags:
    get:
      operationId: listAdminTags
      tags: [admin]
      summary: Tags (vue administration)
      responses:
        "200":
          description: Tags.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tag" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Jeton opaque (pas necessairement un JWT dans le mock : `nextId()`).
        Envoye `Authorization: Bearer <token>` par `src/api/client.ts`.
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Curseur opaque de page (issu de `nextCursor`, `null` = derniere page).
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100 }
      description: Taille de page (20 par defaut cote serveur).
    SlugParam:
      name: slug
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Slug" }
...
    UserIdParam:
      name: userId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
    AnnouncementIdParam:
      name: announcementId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
    ReportIdParam:
      name: reportId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
  responses:
    BadRequest:
      description: "`validation_error` — corps ou parametres invalides."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    Unauthorized:
      description: "`unauthorized` — jeton absent ou invalide."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    Forbidden:
      description: "`forbidden` — role insuffisant pour cette action."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    NotFound:
      description: "`not_found` — ressource introuvable."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
...
    UnprocessableEntity:
      description: "`validation_error` — corps syntaxiquement valide mais violant une regle metier (ex. R-PR1)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
  schemas:
...
    AdminStats: { $ref: "./schemas/AdminStats.json" }
    Announcement: { $ref: "./schemas/Announcement.json" }
    AnnouncementCreateInput: { $ref: "./schemas/AnnouncementCreateInput.json" }
    ApiErrorBody: { $ref: "./schemas/ApiErrorBody.json" }
...
    Id: { $ref: "./schemas/Id.json" }
...
    PaginatedCurrentUser: { $ref: "./schemas/PaginatedCurrentUser.json" }
...
    PaginatedProjectSummary: { $ref: "./schemas/PaginatedProjectSummary.json" }
    PaginatedReportSummary: { $ref: "./schemas/PaginatedReportSummary.json" }
...
    Participation: { $ref: "./schemas/Participation.json" }
...
    Project: { $ref: "./schemas/Project.json" }
    ProjectCreateInput: { $ref: "./schemas/ProjectCreateInput.json" }
...
    ProjectUpdateInput: { $ref: "./schemas/ProjectUpdateInput.json" }
...
    Report: { $ref: "./schemas/Report.json" }
...
    Slug: { $ref: "./schemas/Slug.json" }
    Tag: { $ref: "./schemas/Tag.json" }
...
    UserSummary: { $ref: "./schemas/UserSummary.json" }
...
    AnnouncementWithAuthor:
      description: "Annonce + auteur resolu (`announcementWithAuthorSchema`, src/api/announcements.ts)."
      allOf:
        - { $ref: "#/components/schemas/Announcement" }
        - type: object
          properties:
            author: { $ref: "#/components/schemas/UserSummary" }
          required: [author]
...
    ProjectDeleteInput:
      description: "R-PR7 — confirmation nominative (schema anonyme, src/api/projects.ts)."
      type: object
      properties:
        confirmTitle:
          type: string
          description: Doit correspondre exactement au titre courant du projet.
      required: [confirmTitle]
      additionalProperties: false
...
    AdminReasonInput:
      description: "Motif optionnel d'une action d'administration (schema anonyme, src/api/admin.ts). Recommande de le rendre obligatoire cote backend pour R-S3 (motif toujours notifie) — voir SPEC.md."
      type: object
      properties:
        reason: { type: string, maxLength: 1000 }
      additionalProperties: false